VlasovPoissonTwoSpecies

Modèles cinétiques pour les plasmas de bord

Description du projet

L'interaction d’un plasma avec un bord matériel est un problème très étudié en physique. Lorsqu’un plasma interagit avec une paroi, une couche mince (appelée ”gaine de Debye”) se forme près du bord due au déplacement rapide des électrons (plus léger que les ions) qui sont absorbés par la paroi, créant un déséquilibre de charge près du bord et ainsi la formation d’une couche limite pour le potentiel électrique solution de l’équation de Poisson. La modélisation de ces phénomènes requiert la prise en compte d’équations dites cinétiques pour chaque espèce (électrons et ions). Dans ce cadre, une théorie des états d’équilibre a été établie récemment dans le cas collisionnel [1] et une exploration numérique du cas dynamique a été effectuée dans [2].

L’objectif de ce projet est d’étendre ces travaux récents en prenant en compte les effets collisionnels ou de type source volumique au modèle double espèce (ions et électrons) de Vlasov-Poisson, ce qui s’avère nécessaire pour modéliser les cas pertinents d’un point de vue de la physique. Le cœur du projet sera de construire et d’analyser des méthodes numériques adaptées pour étudier la dynamique de la formation de gaine en prenant en compte les effets collisionnels. Après avoir construit les états stationnaires, l’objectif est de construire des méthodes numériques adaptées pour étudier la dynamique des gaines autour de ces états stationnaires. Plusieurs défis devront être relevés:

capturer les effets multi-échelles présents dans le système,

construire proprement les conditions aux limites pour le potentiel électrique solution de l’équation de Poisson,

définir proprement les schémas numériques d’ordre élevé au bord,

étudier la stabilité des états stationnaires.

Ainsi, des schémas numériques d’ordre élevé préservant l’asymptotique quasi-neutre seront développés dans ce contexte pour pouvoir utiliser des paramètres numériques indépendants des paramètres physiques (comme la longueur de Debye qui correspond à la taille de la couche limite ou le rapport massique entre la masse des électrons et celle des ions) et ainsi réduire le coût des simulations. Ce travail sera aussi l’occasion de faire le point sur les techniques mises en place dans la littérature (physique essentiellement) et de se positionner par rapport à ces travaux.

Références

[1] M. Badsi, M. Campos-Pinto, B. Despre ́s, A minimization formulation of a bi-kinetic

sheath, 2016, Kinetic Related Models.

[2] M. Badsi, M. Merhenberger, L. Navoret, Numerical stability of plasma sheath, 2018, ESAIM Proc.

Le projet est encadré par

Mehdi BADSI, Nantes Université,

Anaïs CRESTETTO, Nantes Université,

— Nicolas CROUSEILLES, Inria Rennes - Bretagne Atlantique — Michel MEHRENBERGER, Aix-Marseille Université.

Le projet est financé par :

ANR MUFFIN à hauteur de 5000 €

— Contrat de recherche à hauteur de 5000 €, — IRMAR Rennes et école doctorale à hauteur de 2000€.

Code

VlasovPoissonTwoSpecies.T_f

VlasovPoissonTwoSpecies.bspline

VlasovPoissonTwoSpecies.compute_e

VlasovPoissonTwoSpecies.compute_energy

VlasovPoissonTwoSpecies.compute_rho

VlasovPoissonTwoSpecies.get_normalized_energy

VlasovPoissonTwoSpecies.landau

VlasovPoissonTwoSpecies.T_f — Function

Return (v*dx+e*dv) f

VlasovPoissonTwoSpecies.bspline — Method

bspline(p, j, x)

Return the value at x in [0,1] of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the De Boor's Algorithm

\[B_{i,0}(x) := \left\{ \begin{matrix} 1 & \mathrm{if} \quad t_i ≤ x < t_{i+1} \\ 0 & \mathrm{otherwise} \end{matrix} \right.\]

\[B_{i,p}(x) := \frac{x - t_i}{t_{i+p} - t_i} B_{i,p-1}(x) + \frac{t_{i+p+1} - x}{t_{i+p+1} - t_{i+1}} B_{i+1,p-1}(x).\]

VlasovPoissonTwoSpecies.compute_e — Method

compute_e(meshx, rho)

compute Ex using that -ik*Ex = rho

VlasovPoissonTwoSpecies.compute_energy — Method

Return the energy e_f = int v^2 f dv

VlasovPoissonTwoSpecies.compute_rho — Method

compute_rho(meshv, f)

Compute charge density ρ(x,t) = ∫ f(x,v,t) dv

VlasovPoissonTwoSpecies.get_normalized_energy — Method

Return the normalized energy |e_f - e_eq| / e_eq
where e_f is the energy of f.

VlasovPoissonTwoSpecies.landau — Method

Landau Damping

Landau damping - Wikipedia